day 10 part 2 improved
Works for test data at least.  May be correct for full data, may not
(it might come up with non-optimal answers in some cases).
Pretty slow, but perhaps not fatally so.
Print statements still in but commented out.

Completes up to #25 in a few hours, but sticks on #26 for more than that.